syncer: move remaining funcs from controllers.go to helpers.go
